Issue #10 Add iPhoneHandTracker to Assets.
This code talks over internet to an iPhone that is running the
application SensorLog over internet. See comments in the .cs file to
get hints about how it should work. Note that it does not necessarily
put the angles into the right places nor does it necessarily use the
right reference frame. This needs some more work.